add Tests

+++ src/main/java/ru/spbu/apcyb/svp/tasks/task3/Task3.java
@@ -1,6 +1,7 @@
-package ru.spbu.apcyb.svp.tasks;
+package ru.spbu.apcyb.svp.tasks.task3;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Arrays;
@@ -13,6 +14,9 @@ public class Task3 {
   private static void getFileStructure(String dirPath, int level, FileWriter out)
       throws IOException {
     File dir = new File(dirPath);
+    if (!dir.exists()) {
+      throw new IOException("No such File or Directory: " + dir);
+    }
     String[] indentation = new String[level];
     Arrays.fill(indentation, "    ");
     out.write(
@@ -32,6 +36,10 @@ public class Task3 {
   public static void main(String[] args) throws IOException {
     String dirPath = args[0];
     String outPath = args[1];
+    File outCheck = new File(outPath);
+    if (!outCheck.exists()) {
+      throw new FileNotFoundException("Incorrect Path: " + outPath);
+    }
     FileWriter out = new FileWriter(outPath);
     getFileStructure(dirPath, 0, out);
     out.close();
